Remove Line type and all associated functions. (#34)
* Remove Line type and all associated functions.

In short, it seems like the Line type is a huge distraction over just
a plain Ray. Most rendering applications for the idea use Rays, and
you can always encode a line as two oppositely facing Rays. I
considered encoding an additional "Mode" into each Ray to describe
whether to interpret it as one-sided or two-sided, but even that
seemed to introduce a lot of complexity which most uses won't need.
If there are workflows that really to benefit from having a
dedicated Line type, I'll happily figure out how to add that back in.

* Remove abandoned RayMode enum